refactor: organize file structure for SignUp
