+++ client/components/page/SignUp/SignUp.page.tsx
@@ -1,0 +1,8 @@
+import { SignUpNextPageProps } from 'pages/signup';
+import SignUp from './SignUp';
+
+const SignUpPage = (props: SignUpNextPageProps) => {
+  return <SignUp {...props} />;
+};
+
+export default SignUpPage;
