+++ client/components/page/SignUp/index.ts
@@ -1,0 +1,1 @@
+export { default as SignUpPage } from './SignUp.page';
